+++ src/Rest/RemoveAllFormSubmissionsController.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Dbp\Relay\FormalizeBundle\Rest;
 
 use Dbp\Relay\CoreBundle\Exception\ApiError;
+use Dbp\Relay\CoreBundle\Rest\CustomControllerTrait;
 use Dbp\Relay\FormalizeBundle\Authorization\AuthorizationService;
 use Dbp\Relay\FormalizeBundle\Service\FormalizeService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -13,6 +14,8 @@ use Symfony\Component\HttpFoundation\Response;
 
 class RemoveAllFormSubmissionsController extends AbstractController
 {
+    use CustomControllerTrait;
+
     public function __construct(
         protected readonly FormalizeService $formalizeService,
         protected readonly AuthorizationService $authorizationService)
@@ -21,9 +24,7 @@ class RemoveAllFormSubmissionsController extends AbstractController
 
     public function __invoke(Request $request): void
     {
-        if (!$this->authorizationService->isAuthenticated()) {
-            throw new ApiError(Response::HTTP_UNAUTHORIZED, 'access denied');
-        }
+        $this->requireAuthentication();
 
         $filters = $request->query->all();
         $form = $this->formalizeService->getForm(Common::getFormIdentifier($filters));
@@ -31,6 +32,8 @@ class RemoveAllFormSubmissionsController extends AbstractController
         if (!$this->authorizationService->isCurrentUserAuthorizedToDeleteFormSubmissions($form)) {
             throw ApiError::withDetails(Response::HTTP_FORBIDDEN, 'forbidden');
         }
+
+        // TODO: consider adding an option to also delete drafts? very dangerous though...
 
         $this->formalizeService->removeAllSubmittedFormSubmissions($form->getIdentifier());
     }